Add/Change #195 (remove the inner loop for doChoice)
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 coli-mpich2.sh.Ray & 83 & 4616837 & 55624 & 126519 &  252372 &  0.9897 & 3 & 30 & 8 \\

+++ code/ExtensionData.h
@@ -102,10 +102,10 @@ public:
 	// reads to check (the ones "in range")
 	bool m_EXTENSION_singleEndResolution;
 	set<uint64_t>::iterator m_EXTENSION_readIterator;
-	map<int,vector<int> > m_EXTENSION_readPositionsForVertices;
-	map<int,vector<int> > m_EXTENSION_pairedReadPositionsForVertices;
-	map<int,vector<int> > m_EXTENSION_pairedLibrariesForVertices;
-	map<int,vector<uint64_t> > m_EXTENSION_pairedReadsForVertices;
+	map<uint64_t,vector<int> > m_EXTENSION_readPositionsForVertices;
+	map<uint64_t,vector<int> > m_EXTENSION_pairedReadPositionsForVertices;
+	map<uint64_t,vector<int> > m_EXTENSION_pairedLibrariesForVertices;
+	map<uint64_t,vector<uint64_t> > m_EXTENSION_pairedReadsForVertices;
 	int m_currentCoverage;
 
 	set<uint64_t>*m_EXTENSION_readsInRange;
